:sparkles: Set cwd for `kai-rpc-server` to a `.vscode/` path (#154)
The `kai-rpc-server` processes generate some working directories that
are created in the process's given working directory. Set the process's
current working directory to a path under the workspace's `.vscode/`
directory. This will ensure that if multiple vscode windows are open,
the konveyor runtime binaries will not clash with each other.

The process will run with a cwd of:
   `{{workspace_root}}/.vscode/konveyor/kai-runtime`

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>