:sparkles: Add support for markdown and json solution progress messages (#372)
Resolves: #352
Supports: https://github.com/konveyor/kai/pull/652
Built on: #374

Add `MarkdownChatMessage` and `JsonChatMessage` rendering to the
resolution webview.

Markdown is configured to support:
  - Basic Markdown [1]
  - GitHub Flavored Markdown [2], [3]
  - HTML embedded in Markdown [4]
- Syntax highlighting [5] in code \`\`\` blocks similar to what GitHub
supports
- Final string renders are sanitized so only "valid" HTML will make it
to the page while invalid HTML is escaped and will show as normal text
[6]

Json messages are formatted, wrapped in a Markdown code block, and
syntax highlighting is used to render it as a message.

[1]: https://remarkjs.github.io/react-markdown
[2]: https://github.github.com/gfm
[3]: https://github.com/remarkjs/react-markdown
[4]: https://github.com/rehypejs/rehype-raw
[5]: https://github.com/rehypejs/rehype-highlight
[6]: https://github.com/rehypejs/rehype-sanitize

---------

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>
Signed-off-by: Ian Bolton <ibolton@redhat.com>
Co-authored-by: Ian Bolton <ibolton@redhat.com>