:bug: Only run partial analysis if a file was changed (#413)
Resolves: #344

Only send a document that has been saved to partial analysis if the document
was actually changed. Just hooking the `onDidSaveTextDocument()` event is
not enough to determine if a change was actually made to a document being
saved. Track document changes by file uri and check that list on save.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>