:bug: Do rpc server actions only if the server is ready (#414)
Resolves: #251

If analyze or getSolution is called, and the server is not running, do
nothing and report it to the output channel. This allows some actions to
become no-ops if the kai jsonrpc server is not running.

The `serverState` changes have been further cleaned up to allow better
state checking in the various action functions.

Previously the checks were not robust enough and used error messages to
report the issue. Error messages here are misleading at best.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>