:bug: Redirect user to analysis page at the end of the konveyor walkthrough (#433)
- Addresses the comment from 
[@Pranav](https://redhat-internal.slack.com/team/UJVF3G38W) on
https://github.com/konveyor/editor-extensions/issues/425 - the
walkthrough button isn’t stateful so we were able to start the server
but there were no notifications or button state on the walkthrough that
let us know that the server was starting/initializing. You could click
it a bunch of times and kick off several instances of the server. The
fix in PR 433 just ushers the user to the analysis page and encourages
the user to start the server from there.


<img width="1447" alt="Screenshot 2025-03-07 at 2 31 35 PM"
src="https://github.com/user-attachments/assets/ca975d14-6e10-4911-b2fb-f0cb2e40fade"
/>

---------

Signed-off-by: Ian Bolton <ibolton@redhat.com>