:bug: Adding helper methods to the proxy, to convert from vscode objects (#1004)
When the proxy makes the vscode commands to proxy the requests return
vscode objects, not the LSP protocol versions. We have to convert them.
There are helpers, here:

https://github.com/microsoft/vscode-languageserver-node/blob/cd32d3e505e8ec8df660bc8e328d5f650673c2d8/client/src/common/codeConverter.ts#L936-L943
but I can not figure out how to get that converter.

If you know how, feel free to either fix in this PR or as a follow up.

<!--
## PR Title Prefix

Every **PR Title** should be prefixed with :text: to indicate its type.

- Breaking change: :warning: (`:warning:`)
- Non-breaking feature: :sparkles: (`:sparkles:`)
- Patch fix: :bug: (`:bug:`)
- Docs: :book: (`:book:`)
- Infra/Tests/Other: :seedling: (`:seedling:`)
- No release note: :ghost: (`:ghost:`)

For example, a pull request containing breaking changes might look like
`:warning: My pull request contains breaking changes`.

Since GitHub supports emoji aliases (ie. `:ghost:`), there is no need to
include
the emoji directly in the PR title -- **please use the alias**. It used
to be
the case that projects using emojis for PR typing had to include the
emoji
directly because GitHub didn't render the alias. Given that `:warning:`
is
easy enough to read as text, easy to parse in release tooling, and
rendered in
GitHub well, we prefer to standardize on the alias.

For more information, please see the Konveyor
[Versioning
Doc](https://github.com/konveyor/release-tools/blob/main/VERSIONING.md).
-->


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

* **New Features**
* Enhanced language server protocol support for improved code
navigation, workspace symbol handling, and better integration with
reference and definition lookup features.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

Signed-off-by: Shawn Hurley <shawn@hurley.page>